chore(RM): move DeviceRemover logic
Moves the logic of DeviceRemover task to a `Core` module so that it can
be tested in isolation w.r.t the creation/actuation of the
`DeviceRemover` task.
